// frontend/src/types/elements.ts
export interface Elements {
    typeSelect: HTMLSelectElement | null;
    categorySelect: HTMLSelectElement | null;
    amountInput: HTMLInputElement | null;
    dateInput: HTMLInputElement | null;
    commentInput: HTMLInputElement | null;
    createButton: HTMLButtonElement | null;
    cancelButton: HTMLButtonElement | null;
    errorContainer: HTMLElement | null;
}